refactor(gateway): extract extensions + jobs + settings + routines slices — ironclaw#2599 stages 4d + 5
Bundles the last feature-slice migrations in one PR. After this lands,
server.rs has zero feature handlers — it's a pure backward-compat
re-export shim that stage 6 will delete.

New feature slices:

- features/extensions/ (stage 4d) — nine routes: list / readiness /
  tools / install / activate / remove / registry / setup / setup-submit.
  Owns derive_activation_status, derive_onboarding, extension_phase_for_web,
  and apply_extension_readiness_to_response. Every handler that takes
  `{name}` from the URL validates via `ExtensionName::new` at the
  boundary (400 on path-traversal / invalid chars / oversized). The
  setup-submit path routes through `AuthManager::resolve_auth_flow_extension_name`
  (canonical resolver) and `platform::engine_dispatch`, preserving the
  identity invariants called out in CLAUDE.md.
- features/jobs/ (stage 5) — nine routes covering sandbox-job lifecycle
  (list / summary / detail / cancel / restart / prompt / events /
  files-list / files-read). Straight file move from handlers/jobs.rs.
- features/settings/ (stage 5) — eight routes (list / export / import /
  get / set / delete / tools-list / tools-set). `resolve_settings_store`
  promoted to `pub(crate)` for `handlers/tool_policy.rs`. Straight file
  move from handlers/settings.rs.
- features/routines/ (stage 5) — seven routes merged from two sources:
  `handlers/routines.rs` (list / summary / detail / trigger / toggle /
  delete) plus the previously-canonical `routines_runs_handler` from
  `server.rs` (the `handlers/routines.rs` copy of the same function was
  marked `#[allow(dead_code)]` and kept in sync manually — that
  redundancy is gone now). Uses the cleaner `routine.is_owned_by(...)`
  ownership predicate throughout instead of the direct `user_id` match
  the `server.rs` copy used.

server.rs changes:

- Deleted 549 lines of extension handlers + 47 lines of
  routines_runs_handler + the now-redundant axum/Json/types/Uuid imports
  the moved handlers pulled in.
- Reduced to 14 lines of `pub use` re-exports for `start_server` and
  `platform::state::*` so external callers (src/main.rs, src/app.rs,
  tests) keep resolving. Stage 6 follow-up deletes even those once the
  callers flip to `platform::*` directly.
- The `#[cfg(test)] mod tests` block stays in place (it's the only
  module still using `test_gateway_state*` helpers that construct state
  for chat + extensions + oauth + pairing together). Test imports
  updated to pull handlers/helpers from their new feature-slice homes.
  Promoting `test_gateway_state*` to `test_helpers.rs` is the last
  blocker before stage 6 and is tracked in the ironclaw#2599 follow-ups
  comment.

Cross-cutting updates:

- platform/router.rs: consolidated feature-slice imports into one
  section with an updated docstring noting the new completion state.
- handlers/mod.rs: dropped `pub mod extensions / jobs / routines /
  settings` declarations (the files physically moved via `git mv`).
  handlers/ now lists only the still-transitional modules.
- handlers/tool_policy.rs: switched `use super::settings::resolve_settings_store`
  to `use crate::channels::web::features::settings::resolve_settings_store`.
- src/extensions/manager.rs: two `handlers::extensions::derive_onboarding`
  call sites redirected to the new slice path.
- src/channels/web/tests/multi_tenant.rs: jobs + routines handler
  imports updated.
- features/{jobs,settings,routines,extensions}/mod.rs: all
  `crate::channels::web::server::{GatewayState, PerUserRateLimiter,
  RateLimiter, ActiveConfigSnapshot}` imports redirected to
  `platform::state::*` directly, so the new slices don't depend on the
  dying shim.

Quality gate:
- cargo fmt clean
- cargo clippy --all --tests --examples --all-features clean
- cargo test --lib channels::web — 424 passed
- scripts/check_gateway_boundaries.py — clean, empty allowlist preserved
- scripts/check_gateway_boundaries.py test — 16/16

Net shape: four slices added under features/, four handlers files
deleted (via `git mv`), server.rs −651 lines (to 14), plus 467 lines
added to features/extensions/ (the nine canonical extension handlers +
helpers). Migration is functionally complete — stage 6 is the final
cleanup.

Explicit non-scope:
- Stage 6 (server.rs shim deletion) is NOT in this PR. Needs the
  `test_gateway_state*` test-helper promotion to `test_helpers.rs` as
  the prerequisite, which is a coordinated cleanup touching every
  caller that still imports `crate::channels::web::server::*`.
- `handlers/` still holds 12 transitional modules (auth, engine,
  frontend, llm, memory, secrets, skills, system_prompt, tokens,
  tool_policy, users, webhooks). Per ironclaw#2599 stage 7, those
  migrate only if churn justifies — most are low-churn and pattern-
  free.
- No `Deps`-view narrowing in any slice; every handler still takes the
  full `GatewayState`. Separate hardening PR per the tracking issue.
