+++ src/channels/web/handlers/mod.rs
@@ -4,10 +4,8 @@
 
 pub mod auth;
 pub mod engine;
-pub mod jobs;
 pub mod llm;
 pub mod memory;
-pub mod routines;
 pub mod secrets;
 pub mod skills;
 pub mod system_prompt;
@@ -15,13 +13,7 @@ pub mod tokens;
 pub mod tool_policy;
 pub mod users;
 
-// Modules not yet wired into server.rs router -- suppress dead_code until
-// they replace their inline counterparts.
-#[allow(dead_code)]
-pub mod extensions;
 pub mod frontend;
-#[allow(dead_code)]
-pub mod settings;
 #[allow(dead_code)]
 pub mod static_files;
 pub mod webhooks;
