+++ src/channels/web/handlers/tool_policy.rs
@@ -14,8 +14,8 @@ use std::sync::Arc;
 
 use axum::{Json, extract::State, http::StatusCode};
 
-use super::settings::resolve_settings_store;
 use crate::channels::web::auth::AdminUser;
+use crate::channels::web::features::settings::resolve_settings_store;
 use crate::channels::web::server::GatewayState;
 use crate::tools::permissions::{
     ADMIN_SETTINGS_USER_ID, ADMIN_TOOL_POLICY_KEY, AdminToolPolicy, parse_admin_tool_policy,
