+++ src/channels/web/server.rs
@@ -1,31 +1,11 @@
-//! Feature handlers for the web gateway that have not yet migrated to
-//! domain modules under `handlers/` or `features/<slice>/`.
+//! Backward-compatibility shim for the ironclaw#2599 migration.
 //!
-//! The platform-level pieces (shared state, rate limiters, CSP/static
-//! serving, route composition, `start_server`) live under
-//! `crate::channels::web::platform::*`. This file re-exports them so
-//! existing `crate::channels::web::server::*` call sites continue to
-//! resolve while the ironclaw#2599 migration is in progress.
+//! All feature handlers have moved to `crate::channels::web::features::<slice>/`.
+//! This module now only re-exports platform types so external callers
+//! (`src/main.rs`, `src/app.rs`, integration tests) keep resolving
+//! `crate::channels::web::server::*`. Stage 6 deletes this file entirely
+//! after those callers flip to `platform::*`.
 
-use std::sync::Arc;
-
-use axum::{
-    Json,
-    extract::{Path, Query, State},
-    http::StatusCode,
-};
-use uuid::Uuid;
-
-use crate::channels::web::auth::AuthenticatedUser;
-use crate::channels::web::types::*;
-
-// --- Backward-compat re-exports for the ironclaw#2599 migration ---
-//
-// The platform-level state types, rate limiters, and `start_server` moved
-// to `crate::channels::web::platform::*`. External callers (handlers,
-// integration tests, `src/main.rs`, `src/app.rs`) still reach them via
-// `crate::channels::web::server::*`; re-export until follow-up PRs update
-// every call site.
 pub use crate::channels::web::platform::router::start_server;
 pub(crate) use crate::channels::web::platform::state::rate_limit_key_from_headers;
 pub use crate::channels::web::platform::state::{
@@ -33,614 +13,15 @@ pub use crate::channels::web::platform::state::{
     PromptQueue, RateLimiter, RoutineEngineSlot, WorkspacePool,
 };
 
-// Job handlers moved to handlers/jobs.rs
-// Logs handlers moved to features/logs/
-
-// --- Extension handlers ---
-
-pub(crate) async fn extensions_list_handler(
-    State(state): State<Arc<GatewayState>>,
-    AuthenticatedUser(user): AuthenticatedUser,
-) -> Result<Json<ExtensionListResponse>, (StatusCode, String)> {
-    let ext_mgr = state.extension_manager.as_ref().ok_or((
-        StatusCode::NOT_IMPLEMENTED,
-        "Extension manager not available (secrets store required)".to_string(),
-    ))?;
-
-    let installed = ext_mgr
-        .list(None, false, &user.user_id)
-        .await
-        .map_err(|e| (StatusCode::INTERNAL_SERVER_ERROR, e.to_string()))?;
-
-    let mut owner_bound_channels = std::collections::HashSet::new();
-    let mut paired_channels = std::collections::HashSet::new();
-    for ext in &installed {
-        if ext.kind == crate::extensions::ExtensionKind::WasmChannel {
-            if ext_mgr.has_wasm_channel_owner_binding(&ext.name).await {
-                owner_bound_channels.insert(ext.name.clone());
-            }
-            if ext_mgr.has_wasm_channel_pairing(&ext.name).await {
-                paired_channels.insert(ext.name.clone());
-            }
-        }
-    }
-    let extensions = installed
-        .into_iter()
-        .map(|ext| {
-            let activation_status =
-                crate::channels::web::handlers::extensions::derive_activation_status(
-                    &ext,
-                    paired_channels.contains(&ext.name),
-                    owner_bound_channels.contains(&ext.name),
-                );
-            let (onboarding_state, onboarding) =
-                crate::channels::web::handlers::extensions::derive_onboarding(
-                    &ext.name,
-                    activation_status,
-                );
-            ExtensionInfo {
-                name: ext.name,
-                display_name: ext.display_name,
-                kind: ext.kind.to_string(),
-                description: ext.description,
-                url: ext.url,
-                authenticated: ext.authenticated,
-                active: ext.active,
-                tools: ext.tools,
-                needs_setup: ext.needs_setup,
-                has_auth: ext.has_auth,
-                activation_status,
-                activation_error: ext.activation_error,
-                version: ext.version,
-                onboarding_state,
-                onboarding,
-            }
-        })
-        .collect();
-
-    Ok(Json(ExtensionListResponse { extensions }))
-}
-
-fn extension_phase_for_web(
-    ext: &crate::extensions::InstalledExtension,
-) -> crate::extensions::ExtensionPhase {
-    if ext.activation_error.is_some() {
-        crate::extensions::ExtensionPhase::Error
-    } else if ext.needs_setup {
-        crate::extensions::ExtensionPhase::NeedsSetup
-    } else if ext.has_auth && !ext.authenticated {
-        crate::extensions::ExtensionPhase::NeedsAuth
-    } else if ext.active
-        || matches!(
-            ext.kind,
-            crate::extensions::ExtensionKind::WasmChannel
-                | crate::extensions::ExtensionKind::ChannelRelay
-        )
-    {
-        crate::extensions::ExtensionPhase::Ready
-    } else {
-        crate::extensions::ExtensionPhase::NeedsActivation
-    }
-}
-
-pub(crate) async fn extensions_readiness_handler(
-    State(state): State<Arc<GatewayState>>,
-    AuthenticatedUser(user): AuthenticatedUser,
-) -> Result<Json<ExtensionReadinessResponse>, (StatusCode, String)> {
-    let ext_mgr = state.extension_manager.as_ref().ok_or((
-        StatusCode::NOT_IMPLEMENTED,
-        "Extension manager not available (secrets store required)".to_string(),
-    ))?;
-
-    let installed = ext_mgr
-        .list(None, false, &user.user_id)
-        .await
-        .map_err(|e| (StatusCode::INTERNAL_SERVER_ERROR, e.to_string()))?;
-
-    let extensions = installed
-        .into_iter()
-        .map(|ext| {
-            let phase = match extension_phase_for_web(&ext) {
-                crate::extensions::ExtensionPhase::Installed => "installed",
-                crate::extensions::ExtensionPhase::NeedsSetup => "needs_setup",
-                crate::extensions::ExtensionPhase::NeedsAuth => "needs_auth",
-                crate::extensions::ExtensionPhase::NeedsActivation => "needs_activation",
-                crate::extensions::ExtensionPhase::Activating => "activating",
-                crate::extensions::ExtensionPhase::Ready => "ready",
-                crate::extensions::ExtensionPhase::Error => "error",
-            }
-            .to_string();
-            ExtensionReadinessInfo {
-                name: ext.name,
-                kind: ext.kind.to_string(),
-                phase,
-                authenticated: ext.authenticated,
-                active: ext.active,
-                activation_error: ext.activation_error,
-            }
-        })
-        .collect();
-
-    Ok(Json(ExtensionReadinessResponse { extensions }))
-}
-
-pub(crate) async fn extensions_tools_handler(
-    State(state): State<Arc<GatewayState>>,
-    AuthenticatedUser(_user): AuthenticatedUser,
-) -> Result<Json<ToolListResponse>, (StatusCode, String)> {
-    let registry = state.tool_registry.as_ref().ok_or((
-        StatusCode::SERVICE_UNAVAILABLE,
-        "Tool registry not available".to_string(),
-    ))?;
-
-    let definitions = registry.tool_definitions().await;
-    let tools = definitions
-        .into_iter()
-        .map(|td| ToolInfo {
-            name: td.name,
-            description: td.description,
-        })
-        .collect();
-
-    Ok(Json(ToolListResponse { tools }))
-}
-
-pub(crate) async fn extensions_install_handler(
-    State(state): State<Arc<GatewayState>>,
-    AuthenticatedUser(user): AuthenticatedUser,
-    Json(req): Json<InstallExtensionRequest>,
-) -> Result<Json<ActionResponse>, (StatusCode, String)> {
-    // When extension manager isn't available, check registry entries for a helpful message
-    let Some(ext_mgr) = state.extension_manager.as_ref() else {
-        // Look up the entry in the catalog to give a specific error
-        if let Some(entry) = state.registry_entries.iter().find(|e| e.name == req.name) {
-            let msg = match &entry.source {
-                crate::extensions::ExtensionSource::WasmBuildable { .. } => {
-                    format!(
-                        "'{}' requires building from source. \
-                         Run `ironclaw registry install {}` from the CLI.",
-                        req.name, req.name
-                    )
-                }
-                _ => format!(
-                    "Extension manager not available (secrets store required). \
-                     Configure DATABASE_URL or a secrets backend to enable installation of '{}'.",
-                    req.name
-                ),
-            };
-            return Ok(Json(ActionResponse::fail(msg)));
-        }
-        return Ok(Json(ActionResponse::fail(
-            "Extension manager not available (secrets store required)".to_string(),
-        )));
-    };
-
-    let kind_hint = req.kind.as_deref().and_then(|k| match k {
-        "mcp_server" => Some(crate::extensions::ExtensionKind::McpServer),
-        "wasm_tool" => Some(crate::extensions::ExtensionKind::WasmTool),
-        "wasm_channel" => Some(crate::extensions::ExtensionKind::WasmChannel),
-        "acp_agent" => Some(crate::extensions::ExtensionKind::AcpAgent),
-        _ => None,
-    });
-
-    match ext_mgr
-        .install(&req.name, req.url.as_deref(), kind_hint, &user.user_id)
-        .await
-    {
-        Ok(result) => {
-            let mut resp = ActionResponse::ok(result.message);
-            match ext_mgr
-                .ensure_extension_ready(
-                    &req.name,
-                    &user.user_id,
-                    crate::extensions::EnsureReadyIntent::PostInstall,
-                )
-                .await
-            {
-                Ok(readiness) => apply_extension_readiness_to_response(&mut resp, readiness, true),
-                Err(e) => {
-                    tracing::debug!(
-                        extension = %req.name,
-                        error = %e,
-                        "Post-install readiness follow-through failed"
-                    );
-                }
-            }
-
-            Ok(Json(resp))
-        }
-        Err(e) => Ok(Json(ActionResponse::fail(e.to_string()))),
-    }
-}
-
-fn apply_extension_readiness_to_response(
-    resp: &mut ActionResponse,
-    readiness: crate::extensions::EnsureReadyOutcome,
-    preserve_success: bool,
-) {
-    match readiness {
-        crate::extensions::EnsureReadyOutcome::Ready { activation, .. } => {
-            if let Some(activation) = activation {
-                resp.message = activation.message;
-                resp.activated = Some(true);
-            }
-        }
-        crate::extensions::EnsureReadyOutcome::NeedsAuth { auth, .. } => {
-            let fallback = format!("'{}' requires authentication.", auth.name);
-            if !preserve_success {
-                resp.success = false;
-                resp.message = auth
-                    .instructions()
-                    .map(String::from)
-                    .unwrap_or_else(|| fallback.clone());
-            } else if let Some(instructions) = auth.instructions() {
-                resp.message = format!("{}. {}", resp.message, instructions);
-            }
-            resp.auth_url = auth.auth_url().map(String::from);
-            resp.awaiting_token = Some(auth.is_awaiting_token());
-            resp.instructions = auth.instructions().map(String::from);
-        }
-        crate::extensions::EnsureReadyOutcome::NeedsSetup {
-            instructions,
-            setup_url,
-            ..
-        } => {
-            if !preserve_success {
-                resp.success = false;
-                resp.message = instructions.clone();
-            } else {
-                resp.message = format!("{}. {}", resp.message, instructions);
-            }
-            resp.instructions = Some(instructions);
-            resp.auth_url = setup_url;
-        }
-    }
-}
-
-pub(crate) async fn extensions_activate_handler(
-    State(state): State<Arc<GatewayState>>,
-    AuthenticatedUser(user): AuthenticatedUser,
-    Path(name): Path<String>,
-) -> Result<Json<ActionResponse>, (StatusCode, String)> {
-    // The URL path segment is user input — validate at the boundary via
-    // `ExtensionName::new` and use the canonical form for all downstream
-    // extension-manager calls and response formatting.
-    let name = ironclaw_common::ExtensionName::new(&name).map_err(|e| {
-        (
-            StatusCode::BAD_REQUEST,
-            format!("Invalid extension name: {e}"),
-        )
-    })?;
-    tracing::trace!(
-        extension = %name.as_str(),
-        user_id = %user.user_id,
-        "extensions_activate_handler: received activate request"
-    );
-    let ext_mgr = state.extension_manager.as_ref().ok_or((
-        StatusCode::NOT_IMPLEMENTED,
-        "Extension manager not available (secrets store required)".to_string(),
-    ))?;
-
-    match ext_mgr
-        .ensure_extension_ready(
-            name.as_str(),
-            &user.user_id,
-            crate::extensions::EnsureReadyIntent::ExplicitActivate,
-        )
-        .await
-    {
-        Ok(readiness) => {
-            let mut resp = ActionResponse::ok(format!("Extension '{}' is ready.", name.as_str()));
-            apply_extension_readiness_to_response(&mut resp, readiness, false);
-            Ok(Json(resp))
-        }
-        Err(err) => Ok(Json(ActionResponse::fail(err.to_string()))),
-    }
-}
-
-pub(crate) async fn extensions_remove_handler(
-    State(state): State<Arc<GatewayState>>,
-    AuthenticatedUser(user): AuthenticatedUser,
-    Path(name): Path<String>,
-) -> Result<Json<ActionResponse>, (StatusCode, String)> {
-    // Validate user-controlled path segment before it reaches the extension
-    // manager — rejects path-traversal, invalid characters, and malformed
-    // slugs with a 400.
-    let name = ironclaw_common::ExtensionName::new(&name).map_err(|e| {
-        (
-            StatusCode::BAD_REQUEST,
-            format!("Invalid extension name: {e}"),
-        )
-    })?;
-    let ext_mgr = state.extension_manager.as_ref().ok_or((
-        StatusCode::NOT_IMPLEMENTED,
-        "Extension manager not available (secrets store required)".to_string(),
-    ))?;
-
-    match ext_mgr.remove(name.as_str(), &user.user_id).await {
-        Ok(message) => Ok(Json(ActionResponse::ok(message))),
-        Err(e) => Ok(Json(ActionResponse::fail(e.to_string()))),
-    }
-}
-
-pub(crate) async fn extensions_registry_handler(
-    State(state): State<Arc<GatewayState>>,
-    AuthenticatedUser(user): AuthenticatedUser,
-    Query(params): Query<RegistrySearchQuery>,
-) -> Json<RegistrySearchResponse> {
-    let query = params.query.unwrap_or_default();
-    let query_lower = query.to_lowercase();
-    let tokens: Vec<&str> = query_lower.split_whitespace().collect();
-
-    // Filter registry entries by query (or return all if empty)
-    let matching: Vec<&crate::extensions::RegistryEntry> = if tokens.is_empty() {
-        state.registry_entries.iter().collect()
-    } else {
-        state
-            .registry_entries
-            .iter()
-            .filter(|e| {
-                let name = e.name.to_lowercase();
-                let display = e.display_name.to_lowercase();
-                let desc = e.description.to_lowercase();
-                tokens.iter().any(|t| {
-                    name.contains(t)
-                        || display.contains(t)
-                        || desc.contains(t)
-                        || e.keywords.iter().any(|k| k.to_lowercase().contains(t))
-                })
-            })
-            .collect()
-    };
-
-    // Cross-reference with installed extensions by (name, kind) to avoid
-    // false positives when the same name exists as different kinds.
-    let installed: std::collections::HashSet<(String, String)> =
-        if let Some(ext_mgr) = state.extension_manager.as_ref() {
-            ext_mgr
-                .list(None, false, &user.user_id)
-                .await
-                .unwrap_or_default()
-                .into_iter()
-                .map(|ext| (ext.name, ext.kind.to_string()))
-                .collect()
-        } else {
-            std::collections::HashSet::new()
-        };
-
-    let entries = matching
-        .into_iter()
-        .map(|e| {
-            let kind_str = e.kind.to_string();
-            RegistryEntryInfo {
-                name: e.name.clone(),
-                display_name: e.display_name.clone(),
-                installed: installed.contains(&(e.name.clone(), kind_str.clone())),
-                kind: kind_str,
-                description: e.description.clone(),
-                keywords: e.keywords.clone(),
-                version: e.version.clone(),
-            }
-        })
-        .collect();
-
-    Json(RegistrySearchResponse { entries })
-}
-
-pub(crate) async fn extensions_setup_handler(
-    State(state): State<Arc<GatewayState>>,
-    AuthenticatedUser(user): AuthenticatedUser,
-    Path(name): Path<String>,
-) -> Result<Json<ExtensionSetupResponse>, (StatusCode, String)> {
-    // Validate user-controlled path segment at entry. Downstream lookups
-    // (`get_setup_schema`, `list().find(...)`) consume the canonical form.
-    let name = ironclaw_common::ExtensionName::new(&name).map_err(|e| {
-        (
-            StatusCode::BAD_REQUEST,
-            format!("Invalid extension name: {e}"),
-        )
-    })?;
-    let ext_mgr = state.extension_manager.as_ref().ok_or((
-        StatusCode::NOT_IMPLEMENTED,
-        "Extension manager not available (secrets store required)".to_string(),
-    ))?;
-
-    let setup = ext_mgr
-        .get_setup_schema(name.as_str(), &user.user_id)
-        .await
-        .map_err(|e| (StatusCode::INTERNAL_SERVER_ERROR, e.to_string()))?;
-
-    let kind = ext_mgr
-        .list(None, false, &user.user_id)
-        .await
-        .ok()
-        .and_then(|list| list.into_iter().find(|e| e.name == name.as_str()))
-        .map(|e| e.kind.to_string())
-        .unwrap_or_default();
-
-    Ok(Json(ExtensionSetupResponse {
-        name: name.as_str().to_string(),
-        kind,
-        secrets: setup.secrets,
-        fields: setup.fields,
-        onboarding_state: None,
-        onboarding: None,
-    }))
-}
-
-pub(crate) async fn extensions_setup_submit_handler(
-    State(state): State<Arc<GatewayState>>,
-    AuthenticatedUser(user): AuthenticatedUser,
-    Path(name): Path<String>,
-    Json(req): Json<ExtensionSetupRequest>,
-) -> Result<Json<ActionResponse>, (StatusCode, String)> {
-    let ext_mgr = state.extension_manager.as_ref().ok_or((
-        StatusCode::NOT_IMPLEMENTED,
-        "Extension manager not available (secrets store required)".to_string(),
-    ))?;
-
-    // The URL path segment is user input — validate at the boundary via
-    // `ExtensionName::new`. Reject path-traversal, invalid characters, or
-    // malformed slugs with a 400 before the value reaches extension
-    // lookup, SSE broadcast, or any `from_trusted` wrap below.
-    let name = ironclaw_common::ExtensionName::new(&name).map_err(|e| {
-        (
-            StatusCode::BAD_REQUEST,
-            format!("Invalid extension name: {e}"),
-        )
-    })?;
-
-    // Clear auth mode regardless of outcome so the next user message goes
-    // through to the LLM instead of being intercepted as a token.
-    crate::channels::web::platform::legacy_auth::clear_auth_mode(&state, &user.user_id).await;
-
-    match ext_mgr
-        .configure(name.as_str(), &req.secrets, &req.fields, &user.user_id)
-        .await
-    {
-        Ok(result) => {
-            // Return ok when activated OR when an OAuth auth_url is present
-            // (activation is expected to be false until OAuth completes).
-            let mut resp = if result.activated || result.auth_url.is_some() {
-                ActionResponse::ok(result.message.clone())
-            } else {
-                ActionResponse::fail(result.message.clone())
-            };
-            resp.activated = Some(result.activated);
-            resp.auth_url = result.auth_url.clone();
-            resp.onboarding_state = result.onboarding_state;
-            resp.onboarding = result.onboarding.clone();
-            let outcome = crate::channels::web::onboarding::classify_configure_result(&result);
-            let mut onboarding_event =
-                crate::channels::web::onboarding::event_from_configure_result(
-                    name.clone(),
-                    &result,
-                    req.thread_id.clone(),
-                );
-            if let (Some(request_id), Some(thread_id)) =
-                (req.request_id.as_deref(), req.thread_id.as_deref())
-            {
-                match outcome {
-                    crate::channels::web::onboarding::ConfigureFlowOutcome::AuthRequired => {}
-                    crate::channels::web::onboarding::ConfigureFlowOutcome::PairingRequired {
-                        instructions,
-                        onboarding,
-                    } => {
-                        let request_id = Uuid::parse_str(request_id).map_err(|_| {
-                            (
-                                StatusCode::BAD_REQUEST,
-                                "Invalid request_id (expected UUID)".to_string(),
-                            )
-                        })?;
-                        if let Some(next_request_id) =
-                            crate::bridge::transition_engine_pending_auth_request_to_pairing(
-                                &user.user_id,
-                                request_id,
-                                Some(thread_id),
-                                name.as_str(),
-                            )
-                            .await
-                            .map_err(|e| (StatusCode::INTERNAL_SERVER_ERROR, e.to_string()))?
-                        {
-                            onboarding_event =
-                                crate::channels::web::types::OnboardingStateDto::pairing_required(
-                                    name.clone(),
-                                    Some(next_request_id),
-                                    Some(thread_id.to_string()),
-                                    Some(result.message.clone()),
-                                    instructions,
-                                    onboarding,
-                                );
-                        }
-                    }
-                    crate::channels::web::onboarding::ConfigureFlowOutcome::Ready => {
-                        crate::channels::web::platform::engine_dispatch::dispatch_engine_external_callback(
-                            &state,
-                            &user.user_id,
-                            thread_id,
-                            request_id,
-                        )
-                        .await?;
-                    }
-                    crate::channels::web::onboarding::ConfigureFlowOutcome::RetryAuth => {}
-                }
-            }
-            // Broadcast the canonical onboarding state so the chat UI can
-            // dismiss or advance any in-progress onboarding UI.
-            state
-                .sse
-                .broadcast_for_user(&user.user_id, onboarding_event);
-            Ok(Json(resp))
-        }
-        Err(e) => {
-            // Preserve the `activated` field on the failure path so clients
-            // (and regression tests) see an explicit `false` rather than
-            // `null`. `ActionResponse::fail` leaves `activated` as `None`,
-            // which serializes to `null` and makes "did activation fail?"
-            // ambiguous from the wire.
-            let mut resp = ActionResponse::fail(e.to_string());
-            resp.activated = Some(false);
-            Ok(Json(resp))
-        }
-    }
-}
-
-// Pairing handlers moved to features/pairing/
-
-pub(crate) async fn routines_runs_handler(
-    State(state): State<Arc<GatewayState>>,
-    AuthenticatedUser(user): AuthenticatedUser,
-    Path(id): Path<String>,
-) -> Result<Json<serde_json::Value>, (StatusCode, String)> {
-    let store = state.store.as_ref().ok_or((
-        StatusCode::SERVICE_UNAVAILABLE,
-        "Database not available".to_string(),
-    ))?;
-
-    let routine_id = Uuid::parse_str(&id)
-        .map_err(|_| (StatusCode::BAD_REQUEST, "Invalid routine ID".to_string()))?;
-
-    // Verify ownership before listing runs.
-    let routine = store
-        .get_routine(routine_id)
-        .await
-        .map_err(|e| (StatusCode::INTERNAL_SERVER_ERROR, e.to_string()))?
-        .ok_or((StatusCode::NOT_FOUND, "Routine not found".to_string()))?;
-
-    if routine.user_id != user.user_id {
-        return Err((StatusCode::NOT_FOUND, "Routine not found".to_string()));
-    }
-
-    let runs = store
-        .list_routine_runs(routine_id, 50)
-        .await
-        .map_err(|e| (StatusCode::INTERNAL_SERVER_ERROR, e.to_string()))?;
-
-    let run_infos: Vec<RoutineRunInfo> = runs
-        .iter()
-        .map(|run| RoutineRunInfo {
-            id: run.id,
-            trigger_type: run.trigger_type.clone(),
-            started_at: run.started_at.to_rfc3339(),
-            completed_at: run.completed_at.map(|dt| dt.to_rfc3339()),
-            status: run.status.to_string(),
-            result_summary: run.result_summary.clone(),
-            tokens_used: run.tokens_used,
-            job_id: run.job_id,
-        })
-        .collect();
-
-    Ok(Json(serde_json::json!({
-        "routine_id": routine_id,
-        "runs": run_infos,
-    })))
-}
-
-// Gateway status handler moved to features/status/
-
 #[cfg(test)]
 mod tests {
     use super::*;
+    use std::sync::Arc;
+
+    use axum::{Json, extract::State, http::StatusCode};
+
+    use uuid::Uuid;
+
     use crate::agent::SessionManager;
     use crate::auth::oauth;
     use crate::channels::relay::DEFAULT_RELAY_NAME;
@@ -649,6 +30,11 @@ mod tests {
         IN_PROGRESS_STALE_AFTER_MINUTES, chat_approval_handler, chat_auth_cancel_handler,
         chat_auth_token_handler, chat_gate_resolve_handler, chat_history_handler,
         pending_gate_extension_name,
+    };
+    use crate::channels::web::features::extensions::{
+        apply_extension_readiness_to_response, extension_phase_for_web,
+        extensions_activate_handler, extensions_list_handler, extensions_readiness_handler,
+        extensions_remove_handler, extensions_setup_handler, extensions_setup_submit_handler,
     };
     use crate::channels::web::features::oauth::{
         oauth_callback_handler, slack_relay_oauth_callback_handler,
@@ -663,6 +49,7 @@ mod tests {
         css_handler, generate_csp_nonce, stamp_nonce_into_html,
     };
     use crate::channels::web::sse::SseManager;
+    use crate::channels::web::types::*;
     use crate::channels::web::types::{
         ExtensionActivationStatus, classify_wasm_channel_activation,
     };
